Titulo: PASAR TESTS EN INTEGRACION CONTINUA Mensaje: Arreglar los tests para hacerlos con integracion continua.
Que queda por hacer: Probar que funciona.

ID: SF-01
